requirements.txt
requests
beautifulsoup4
python-dotenv
python-telegram-bot